src/requirements.txt
flask
jieba
scrapy
kafka-python==2.0.2